todo/requirements.txt
# common
redis==3.2.1
celery==4.3.0
flower==0.9.3
psycopg2-binary==2.8.2

# django
django==2.2.2
django-cors-headers==3.0.2
django-celery-email==2.0.2
django-model-utils==3.1.2
django-debug-toolbar==1.11
django-taggit==1.1.0
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework_simplejwt==4.3.0
djangorestframework-filters==v1.0.0.dev0
drf-dynamic-fields==0.3.1
drf-writable-nested==0.5.1
coverage==4.5.3
freezegun==0.3.12
pytest==4.5.0
pytest-django==3.4.8
pytest-cov==2.7.1
pytest-xdist==1.28.0
pydevd-pycharm~=191.7479.30
supervisor==4.0.3
